Allow specification of a jsonschema.FormatChecker
Also, only validate the provided schemas once, not on every
function call. (We still validate the request and response bodies
on each call, though)

A `jsonschema.FormatChecker` instance, when provided, allows one to
validate the contents of string-type fields. See
http://python-jsonschema.readthedocs.io/en/latest/validate/#validating-formats
for details.

These changes add the ability to specify a `FormatChecker` along with
the schema body at decoration time.

Previously we were using jsonschema.validate() to validate our
request and response bodies. It invokes two validation operations: one
to validate the schema and one to validate the instance against the
schema. We don't need to validate the schema on every request, as that
is redundant work beyond the first call.

These changes make it so that the schema is validated once when being
defined in the @request_schema or @response_schema decorators at
class/method creation time.

The added tests demonstrate how one might depend on external libraries
to check well-known formats. Furthermore, they demonstrate how one might
register a custom format checker as well.